third_party/WebKit/LayoutTests/platform/mac-retina/external/wpt/webvtt/api/VTTCue/text-expected.txt

CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
This is a testharness.js-based test.
FAIL VTTCue.text, script-created cue cue is not defined
FAIL VTTCue.text, parsed cue assert_unreached: got error event Reached unreachable code
Harness: the test ran to completion.
